+++ src/os_v5/utility/argument_parser/single_argument_parsers.ts
@@ -1,41 +1,6 @@
 import { ConsoleUtility } from "shared/utility/console_utility/console_utility"
 import { RoomName } from "shared/utility/room_name_types"
-
-export type ArgumentKey = string | number
-type Key = ArgumentKey
-
-export type ArgumentParserOptions = {
-  readonly errorMessage?: (rawArgument: string, error: string) => string
-}
-
-export abstract class SingleArgumentParser<Options, Value> {
-  public constructor(
-    public readonly key: Key,
-    public readonly value: string | null,
-    protected readonly parseOptions?: ArgumentParserOptions,
-  ) {
-  }
-
-  /** throws */
-  public abstract parse(options?: Options): Value
-
-  protected missingArgumentErrorMessage(): string {
-    return `Missing ${getKeyName(this.key)} argument`
-  }
-}
-
-export abstract class SingleOptionalArgument<Options, Value> extends SingleArgumentParser<Options, Value> {
-  /** throws */
-  public abstract parse(options?: Options): Value
-
-  /** throws */
-  public parseOptional(options?: Options): Value | null {
-    if (this.value == null) {
-      return null
-    }
-    return this.parse(options)
-  }
-}
+import { ArgumentKey, getKeyName, SingleOptionalArgument } from "./single_argument_parser"
 
 
 // ---- Primitive Type ---- //
@@ -74,32 +39,9 @@ export class RoomNameArgument extends SingleOptionalArgument<{ my?: boolean, all
 }
 
 
-// ---- Utility ---- //
-const getKeyName = (key: Key): string => {
-  if (typeof key === "string") {
-    return key
-  }
-  const index = key
-  const indexName = ((): string => {
-    switch (index % 10) {
-    case 1:
-      return "st"
-    case 2:
-      return "nd"
-    case 3:
-      return "rd"
-    default:
-      return "th"
-    }
-  })()
-  return `${index}${indexName}`
-}
-
-
-
 // ---- Parser ---- //
 /** throws */
-const parseIntValue = (key: Key, value: string, options?: { min?: number, max?: number }): number => {
+const parseIntValue = (key: ArgumentKey, value: string, options?: { min?: number, max?: number }): number => {
   const intValue = parseInt(value, 10)
   if (isNaN(intValue) === true) {
     throw `${value} is not an integer value`
@@ -109,7 +51,7 @@ const parseIntValue = (key: Key, value: string, options?: { min?: number, max?: 
 }
 
 /** throws */
-const validateNumberRange = (key: Key, value: number, options?: { min?: number, max?: number }): void => {
+const validateNumberRange = (key: ArgumentKey, value: number, options?: { min?: number, max?: number }): void => {
   if (options?.min != null && value < options.min) {
     throw `${getKeyName(key)} is too small (${value} < ${options.min})`
   }
